Create GoS65
senselessclay has another keyboard that isn't added to the v3 repository, it doesn't work in via without using the design tab to load a draft definition. I have the json, the hex file for the keyboard, as well as the default keymap to add. I just need help doing so, the GoS was an extremely small run of about 30-40 boards and the person that did the group buy isn't around anymore in the community so I'm hoping to be able to make it compatible with VIA even though I am not the maker of the keyboard. I'm sorry if this is the incorrect way to go about doing this, thank you.